about to refactor share modal, just want to commit before everything messes up
